move example and file api into different modules
