+++ connector-aws-s3-example/src/main/resources/application.properties
