+++ connector-aws-s3-example/src/main/resources/bpmn/example.bpmn
