+++ connector-aws-s3/src/main/java/de/novatec/bpm/camunda/connector/aws/s3/adapter/in/process/model/AuthenticationRequestData.java
