+++ connector-aws-s3/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+#zeebe.client.cloud.region: my-region
+#zeebe.client.cloud.clusterid: my-cluster-id
+#zeebe.client.cloud.clientid: my-client-id
+#zeebe.client.cloud.clientsecret: my-client-secret
+
+operate.client.enabled=true
+camunda.operate.client.url=http://localhost:8081
+camunda.operate.client.username=demo
+camunda.operate.client.password=demo
+zeebe.client.broker.gateway-address=localhost:26500
+zeebe.client.security.plaintext=true
