+++ connector-aws-s3/src/main/resources/logback.xml
