+++ connector-aws-s3/src/test/java/de/novatec/bpm/camunda/connector/aws/s3/ConnectorAdapterContextTest.java
