+++ connector-aws-s3/src/test/resources/.env
@@ -1,0 +1,20 @@
+## Image versions ##
+CAMUNDA_CONNECTORS_VERSION=0.18.1
+CAMUNDA_OPTIMIZE_VERSION=3.10.0
+CAMUNDA_PLATFORM_VERSION=8.2.13
+CAMUNDA_WEB_MODELER_VERSION=8.2.5
+ELASTIC_VERSION=7.17.9
+KEYCLOAK_SERVER_VERSION=19.0.3
+MAILPIT_VERSION=v1.5.4
+POSTGRES_VERSION=14.5-alpine
+
+## Configuration ##
+# By default the zeebe api is public, when setting this to `identity` a valid zeebe client token is required
+ZEEBE_AUTHENTICATION_MODE=none
+ZEEBE_CLIENT_ID=zeebe
+ZEEBE_CLIENT_SECRET=zecret
+
+# Set to 'true' to enable resource based authorizations for users and groups
+# This can be used to limit access for users or groups to view/update specific
+# processes and decisions in Operate and Tasklist
+RESOURCE_AUTHORIZATIONS_ENABLED=false
