+++ connector-aws-s3/src/test/resources/invoices/invoice.txt
