+++ connector-file-api/src/main/java/de/novatec/bpm/camunda/connector/file/api/FileCommand.java
@@ -1,0 +1,15 @@
+package de.novatec.bpm.camunda.connector.file.api;
+
+import de.novatec.bpm.camunda.connector.file.api.impl.model.RequestData;
+
+import java.io.IOException;
+
+public interface FileCommand {
+
+    RequestData uploadFile(RequestData request) throws IOException;
+
+    RequestData deleteFile(RequestData request) throws IOException;
+
+    RequestData downloadFile(RequestData request) throws IOException;
+
+}
