+++ connector-file-api/src/main/java/de/novatec/bpm/camunda/connector/file/api/LocalFileCommand.java
@@ -1,0 +1,14 @@
+package de.novatec.bpm.camunda.connector.file.api;
+
+import java.io.IOException;
+import java.nio.file.Path;
+
+public interface LocalFileCommand {
+
+    Path saveFile(byte[] content, String filePath) throws IOException;
+
+    byte[] loadFile(String filePath) throws IOException;
+
+    void deleteFile(String filePath) throws IOException;
+
+}
