+++ connector-file-api/src/main/java/de/novatec/bpm/camunda/connector/file/api/impl/model/FileContent.java
@@ -1,0 +1,14 @@
+package de.novatec.bpm.camunda.connector.file.api.impl.model;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class FileContent {
+
+    private byte[] content;
+    private String contentType;
+    private Long contentLength;
+
+}
