+++ connector-file-api/src/main/java/de/novatec/bpm/camunda/connector/file/api/impl/model/RequestData.java
@@ -1,0 +1,18 @@
+package de.novatec.bpm.camunda.connector.file.api.impl.model;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class RequestData {
+
+    private String accessKey;
+    private String secretKey;
+    private String bucket;
+    private String key;
+    private String region;
+    private String filePath;
+    private String contentType;
+
+}
